+++ components/firstpair/api/src/main/java/com/flipperdevices/firstpair/api/FirstPairFeatureEntry.kt
@@ -1,0 +1,11 @@
+package com.flipperdevices.firstpair.api
+
+import com.flipperdevices.core.ui.navigation.AggregateFeatureEntry
+import com.flipperdevices.core.ui.navigation.FeatureScreenRootRoute
+
+interface FirstPairFeatureEntry : AggregateFeatureEntry {
+    override val ROUTE: FeatureScreenRootRoute
+        get() = FeatureScreenRootRoute.FIRST_PAIR
+
+    fun getFirstPairScreen(): String
+}